Modifies the is_cluster_terminated check to include instances that belong to the placement group
It seems that when a instance is terminated in a VPC the security group is automatically removed
from the instance metadata causing the is_cluster_terminated to return a false positive.
The code will then continue and try to remove the placement group but time out since the
placement group cannot be removed with instances remaining.